# Optimize ULID Compare performance