ULID can be seen as two uint64 types. The previous comparison method performed multiple meaningless operations (such as checking if the lengths are equal). The new method addresses these issues; please refer to the benchmark tests for specifics.

Old:
```
cpu: AMD Ryzen 9 4900H with Radeon Graphics
=== RUN   BenchmarkOld
BenchmarkOld
BenchmarkOld-16
346184070                3.469 ns/op    9225.41 MB/s           0 B/op          0 allocs/op
PASS
```

New:
```
cpu: AMD Ryzen 9 4900H with Radeon Graphics
=== RUN   BenchmarkNew
BenchmarkNew
BenchmarkNew-16
604745437                1.975 ns/op    16203.47 MB/s          0 B/op          0 allocs/op
PASS
```