@Web-Coke 2025-05-16
I am very sorry for the oversight of failing to distinguish between Little Endian mode and Big Endian mode. The code runs fine in Big Endian mode, but the results in Little Endian mode are abnormal! I have now fixed this issue and will resubmit a merge request later, along with the latest test results!